requirements.txt
Flask>=0.10.1
python-telegram-bot>=2.5
numpy
opencv-python
gunicorn
tensorflow
requests